Move some preallocate proptests to their own file
And fix the test module structure so it is consistent with the rest of
zebra-chain.
