+++ zebra-chain/src/serialization/tests.rs
@@ -1,0 +1,2 @@
+mod preallocate;
+mod prop;
